pyproject.toml
[tool.poetry]
name = "game"
version = "0.1.0"
description = ""
authors = ["Szauer Szilárd <szauerszilard@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
tensorflow = "^2.20.0"
numpy = "^2.3.2"
pandas = "^2.3.1"
matplotlib = "^3.10.5"
seaborn = "^0.13.2"
scikit-learn = "^1.7.1"
pygame-ce = "^2.5.5"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pre-commit = "^4.3.0"
ruff = "^0.12.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.poetry.scripts]
game = "main:main"
